Move row transformers to frontend/bq_schemas.py
