+++ openprescribing/ebmdatalab/bigquery_old.py
@@ -23,6 +23,7 @@ from googleapiclient import discovery
 from oauth2client.client import GoogleCredentials
 
 from frontend.bq_schemas import DMD_SCHEMA, CCG_SCHEMA, PRESCRIBING_SCHEMA, PRESENTATION_SCHEMA, PRACTICE_SCHEMA, PRACTICE_STATISTICS_SCHEMA
+from frontend.bq_schemas import prescribing_transform, statistics_transform, presentation_transform, ccgs_transform
 
 
 logger = logging.getLogger(__name__)
@@ -102,45 +103,6 @@ def load_data_from_file(
         return job
 
 
-def prescribing_transform(row):
-    """Transform a row from a formatted file into data suitable for
-    storing in our bigquery schema
-
-
-    A 'formatted file' is a file created by the
-    import_hscic_prescribing Django management command.
-
-    """
-    # To match the prescribing table format in BigQuery, we have
-    # to re-encode the date field as a bigquery TIMESTAMP and drop
-    # a couple of columns
-    row[10] = "%s 00:00:00" % row[10]
-    del(row[3])
-    del(row[-1])
-    return row
-
-
-def statistics_transform(row):
-    """Transform a row from the frontend_practicestatistics table so it
-    matches our statistics schema
-
-    """
-    row[0] = "%s 00:00:00" % row[0]  # BQ TIMESTAMP format
-    return row
-
-
-def presentation_transform(row):
-    """Transform a row from the frontend_presentation table so it
-    matches our statistics schema
-
-    """
-    if row[2] == 't':
-        row[2] = 'true'
-    else:
-        row[2] = 'false'
-    return row
-
-
 def load_prescribing_data_from_file(
         dataset_name, table_name, source_file_name):
     """Given a formatted file of prescribing data, load it into BigQuery.
@@ -181,16 +143,9 @@ def load_ccgs_from_pg(dataset='hscic'):
     application into BigQuery
 
     """
-    def transform(row):
-        if row[4]:
-            row[4] = "%s 00:00:00" % row[4]
-        if row[5]:
-            row[5] = "%s 00:00:00" % row[5]
-        return row
-
     load_data_from_pg(
         dataset, 'ccgs', 'frontend_pct',
-        CCG_SCHEMA, cols=[x.name for x in CCG_SCHEMA], _transform=transform)
+        CCG_SCHEMA, cols=[x.name for x in CCG_SCHEMA], _transform=ccgs_transform)
 
 
 def load_data_from_pg(dataset_name, bq_table_name,
